+++ internal/app/models.go
@@ -1,7 +1,5 @@
 package app
 
-import "github.com/aws/aws-sdk-go/service/secretsmanager"
-
 // TagProtocolValue should match this value in order to indicate that
 // a certain protocol should be managed on tagged security group.
 // In any case, only "owned" rules will be managed.
@@ -18,7 +16,7 @@ type Config struct {
 	Rules     []*Rule              `json:"rules"`
 }
 
-// Protocol represends a single protocol configuration
+// Protocol represents a single protocol configuration
 type Protocol struct {
 	Transport *string `json:"transport"`
 	FromPort  *int64  `json:"from_port"`
@@ -28,9 +26,4 @@ type Protocol struct {
 // Rule represents a whitelisted CIDR
 type Rule struct {
 	CIDR *string `json:"cidr"`
-}
-
-// Client represents a Secrets Manager client
-type Client interface {
-	GetSecretValue(*secretsmanager.GetSecretValueInput) (*secretsmanager.GetSecretValueOutput, error)
 }